requirements.txt
openai
pandas
python-dotenv
tqdm
unbabel-comet
torch
nest_asyncio
